requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
python-dotenv>=1.0.0
resemblyzer>=0.0.1
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.3
scipy>=1.10.1
aiohttp>=3.9.1
pydantic-settings>=2.1.0
aiofiles>=23.2.1

google-generativeai>=0.2.0
openai>=1.3.0
openai-whisper

pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

jinja2>=3.1.2
email-validator>=2.0.0
resend